home page styling, step4, setup
