+++ src/App.js
@@ -1,23 +1,23 @@
-import React, { useEffect, useState } from "react"
-import { Switch, Route } from "react-router-dom"
-import { useDispatch, useSelector } from "react-redux"
+import React, { useEffect, useState } from "react";
+import { Switch, Route } from "react-router-dom";
+import { useDispatch, useSelector } from "react-redux";
 
 //Pages
-import Homepage from "./pages/Homepage"
-import Dashboard from "./pages/Dashboard"
-import Login from "./pages/Login"
-import SignUp from "./pages/SignUp"
-import CodingGame from "./pages/CodingGame"
+import Homepage from "./pages/Homepage/Homepage";
+import Dashboard from "./pages/Dashboard";
+import Login from "./pages/Login";
+import SignUp from "./pages/SignUp";
+import CodingGame from "./pages/CodingGame";
 
 // Components
-import Navbar from "./components/Navbar"
-import Motivation from "./pages/Motivation"
-import TriviaQuiz from "./pages/TriviaQuiz"
+import Navbar from "./components/Navbar";
+import Motivation from "./pages/Motivation";
+import TriviaQuiz from "./pages/TriviaQuiz";
 
 // Mui components
-import { getUserWithStoredToken } from "./store/user/actions"
-import { ThemeProvider, createMuiTheme } from "@material-ui/core/styles"
-import { Paper, CssBaseline, Container } from "@material-ui/core"
+import { getUserWithStoredToken } from "./store/user/actions";
+import { ThemeProvider, createMuiTheme } from "@material-ui/core/styles";
+import { Paper, CssBaseline, Container } from "@material-ui/core";
 
 const lightTheme = createMuiTheme({
   palette: {
@@ -31,7 +31,7 @@ const lightTheme = createMuiTheme({
     type: "light",
     // background: { paper: "#7A9D96" },
   },
-})
+});
 
 const darkTheme = createMuiTheme({
   palette: {
@@ -44,15 +44,15 @@ const darkTheme = createMuiTheme({
     },
     type: "dark",
   },
-})
+});
 
 function App() {
-  const [darkMode, set_darkMode] = useState(true)
-  const dispatch = useDispatch()
+  const [darkMode, set_darkMode] = useState(true);
+  const dispatch = useDispatch();
 
   useEffect(() => {
-    dispatch(getUserWithStoredToken())
-  }, [dispatch])
+    dispatch(getUserWithStoredToken());
+  }, [dispatch]);
 
   return (
     <ThemeProvider theme={darkMode ? darkTheme : lightTheme}>
@@ -70,7 +70,7 @@ function App() {
         </Switch>
       </Container>
     </ThemeProvider>
-  )
+  );
 }
 
-export default App
+export default App;
